# docs/swagger.yaml
definitions:
...
  requests.CreateProductCategoryRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      parent_category_id:
        type: string
    required:
    - name
    type: object
...
  responses.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
...
  responses.ProductCategoryResponse:
    properties:
      child_categories:
        items:
          $ref: '#/definitions/responses.ProductCategoryResponse'
        type: array
      create_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      parent_category:
        $ref: '#/definitions/responses.ProductCategoryResponse'
      update_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the core backend service for SEP490 project with authentication,
    user management, and business features.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SEP490 Core Backend API
  version: "1.0"
paths:
...
  /api/v1/categories:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of product categories. Optional search
        by name.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search term for category name
        in: query
        name: search
        type: string
      - default: false
        description: Include deleted categories (null - get all, true - only deleted,
          false - only active)
        in: query
        name: deleted
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.ProductCategoryResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get product categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new product category
      parameters:
      - description: Create category payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.CreateProductCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.ProductCategoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a product category
      tags:
      - Categories
  /api/v1/categories/{id}:
    patch:
      consumes:
      - application/json
      description: Delete category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a product category
      tags:
      - Categories
  /api/v1/categories/{id}/parent:
    patch:
      consumes:
      - application/json
      description: Set a parent category for an existing category. If parent_id is
        not provided, the parent category will be removed.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Parent category ID
        in: query
        name: parent_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ProductCategoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Assign parent category
      tags:
      - Categories
...
  /api/v1/contract_payments/contract/{contract_id}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Contract retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid request or validation error
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Brand not found
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "409":
          description: Contract number already exists
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Create contract payments based on a contract ID
      tags:
      - Contract Payments
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"